+++ lib/eggnstone_google_crashlytics.dart
@@ -1,4 +1,5 @@
 library eggnstone_google_crashlytics;
 
-export 'package:eggnstone_google_crashlytics/google/GoogleCrashlyticsService.dart';
+export 'package:eggnstone_google_crashlytics/google/GoogleCrashlyticsNonWebService.dart' if (dart.library.html) 'package:eggnstone_google_crashlytics/google/GoogleCrashlyticsWebService.dart';
+//export 'package:eggnstone_google_crashlytics/google/GoogleCrashlyticsWebService.dart';
 export 'package:eggnstone_google_crashlytics/google/IGoogleCrashlyticsService.dart';